// apps/04-news/src/app/interfaces/news-api.interface.ts
export interface NewsResponse {
    status: string;
    totalResults: number;
    articles: Article[];
}

export interface Article {
    source: Source;
    author?: string;
    title: string;
    description: string;
    url: string;
    urlToImage: string;
    publishedAt: Date | string;
    content: string;
}

export interface Source {
    id?: string;
    name: string;
}